Refactor ReminderManager into an interface for testing
