+++ app/src/main/java/com/willbsp/habits/domain/usecase/SaveHabitUseCase.kt
@@ -6,7 +6,7 @@ import com.willbsp.habits.data.repository.ReminderRepository
 import com.willbsp.habits.domain.model.HabitData
 import com.willbsp.habits.domain.model.HabitReminderType
 import com.willbsp.habits.domain.model.toHabit
-import com.willbsp.habits.util.ReminderManager
+import com.willbsp.habits.util.LocalReminderManager
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
@@ -16,7 +16,7 @@ import javax.inject.Inject
 class SaveHabitUseCase(
     private val habitRepository: HabitRepository,
     private val reminderRepository: ReminderRepository,
-    private val reminderManager: ReminderManager,
+    private val reminderManager: LocalReminderManager,
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
 ) {
 
@@ -24,7 +24,7 @@ class SaveHabitUseCase(
     constructor(
         habitRepository: HabitRepository,
         reminderRepository: ReminderRepository,
-        reminderManager: ReminderManager
+        reminderManager: LocalReminderManager
     ) : this(habitRepository, reminderRepository, reminderManager, Dispatchers.IO)
 
     suspend operator fun invoke(data: HabitData, habitId: Int? = null) =
