+++ app/src/main/java/com/willbsp/habits/util/AlarmPermissionReceiver.kt
@@ -10,7 +10,7 @@ import javax.inject.Inject
 class AlarmPermissionReceiver : BroadcastReceiver() {
 
     @Inject
-    lateinit var reminderManager: ReminderManager
+    lateinit var reminderManager: LocalReminderManager
 
     override fun onReceive(context: Context, intent: Intent) = goAsync {
